Framebfufer/Mouse Position Solved
Fly window size was not being updated when framebuffer resized

Actually, Framebuffer and window could have different sizes, but will take care of that later.

To add list:
-Allowing Inside resolution of the framebuffer, allow linkage of window size to framebuffer size, this will be engine option but library should support an easy way to do so.
- Actuall Window and Framebuffer properties should follow a similar system to attributes, that it stores the option in a macro way, then Flylib will have their wrapper function that works per library using those values.

+++ Temp_Engine/FlyLib/flylib_src/fly_input.cpp
@@ -346,7 +346,7 @@ FLYINPUT_Actions FLYINPUT_DetectGesture(FLY_Pointer& p)
     // Evaluate what the pointer has done during the time it was tracked
     int16 delta_x = p.gesture.end_x - p.gesture.start_x;
     int16 delta_y = p.gesture.end_y - p.gesture.start_y;
-    uint16 screen_w, screen_h;
+    int32 screen_w, screen_h;
     FLYDISPLAY_GetWindowSize(0, &screen_w, &screen_h);
 
     uint16 abs_delta_x = delta_x;
